+++ npm-packages/react-api/src/hooks/editor/use-command.ts
@@ -1,5 +1,5 @@
 import {
-  ExtensionCommandInfo,
+  CommandInfo,
   IMCMessage,
   IMCMessageTypeEnum,
   ReceiverHandler,
@@ -10,15 +10,19 @@ import { useEffect, useState } from "react";
 /**
  * Register an extension command to listen to IMC messages from the core,
  * and pass to the extension to handle.
+ *
+ * @param commandInfo Command information to register.
+ * @param initialHandler Initial handler function to handle the command.
+ *
  */
-export default function useExtCommand(
-  commandInfo: ExtensionCommandInfo,
+export default function useCommand(
+  commandInfo: CommandInfo,
   initialHandler?: (args: any) => Promise<string | void>
 ) {
   const { isReady, imc } = useIMC(getReceiverHandlerMap());
 
   const [handler, setHandler] = useState<
-    ((args: any) => Promise<string | void>) | undefined
+    ((args: any) => Promise<any>) | undefined
   >(initialHandler);
 
   useEffect(() => {
@@ -83,10 +87,9 @@ export default function useExtCommand(
 
   /**
    *
-   * @param handler Function to handle the command. Return a string in the handler to let
-   * Pulse Editor assistant to read it out.
+   * @param handler Function to handle the command.
    */
-  function updateHandler(handler: (args: any) => Promise<string | void>) {
+  function updateHandler(handler: (args: any) => Promise<any>) {
     setHandler(() => handler);
   }