+++ web/components/views/standalone-app/standalone-app-view.tsx
@@ -3,15 +3,13 @@ import BaseAppView from "../base/base-app-view";
 import ViewControlLayout from "../layout/view-control-layout";
 
 export default function StandaloneAppView({
-  viewId,
   config,
 }: {
-  viewId: string;
   config: AppViewConfig;
 }) {
   return (
     <ViewControlLayout type="app">
-      <BaseAppView viewId={viewId} config={config} />
+      <BaseAppView viewId={config.viewId} config={config} />
     </ViewControlLayout>
   );
 }